turn doubles to integers.

+++ Point.cpp
@@ -2,11 +2,11 @@
 
 #include <math.h>
 
-Point::Point() : x(0.0), y(0.0), z(0.0)
+Point::Point() : x(0), y(0), z(0)
 {
 }
 
-Point::Point(double x, double y, double z) : x(x), y(y), z(z)
+Point::Point(int x, int y, int z) : x(x), y(y), z(z)
 {
 }
 
@@ -26,12 +26,12 @@ Point & Point::operator=(const Point & point)
 	return *this;
 }
 
-double Point::distanceTo(double x, double y, double z) const
+int Point::distanceTo(int x, int y, int z) const
 {
 	return sqrt((x - this->x)*(x - this->x) + (y - this->y)*(y - this->y) + (z - this->z)*(z - this->z));
 }
 
-double Point::distanceTo(const Point & point) const
+int Point::distanceTo(const Point & point) const
 {
 	return sqrt((point.x - this->x)*(point.x - this->x) + (point.y - this->y)*(point.y - this->y) + (point.z - this->z)*(point.z - this->z));
 }
